2014-09-17 Fred Gleason <fredg@paravelsystems.com> 	* Fixed warnings in 'libp/asterisk_driver.cpp'. 	* Fixed warnings in 'libp/edit_connection_data.cpp'. 	* Fixed warnings in 'libp/escapestring.cpp'. 	* Fixed warnings in 'libp/message_data.cpp'. 	* Fixed warnings in 'libp/ml_timeslot.cpp'. 	* Fixed warnings in 'libp/mlconf.cpp'. 	* Fixed warnings in 'libp/mllabel.cpp'. 	* Fixed warnings in 'libp/mlttydevice.cpp'. 	* Fixed warnings in 'mlhpi/mlplaystream.cpp'. 	* Fixed warnings in 'mlhpi/mlsoundcard.cpp'. 	* Fixed warnings in 'mlhpi/mlrecordstream.cpp'. 	* Fixed warnings in 'mlhpi/mlwavefile.cpp'. 	* Fixed warnings in 'mdbadmin/edit_show.cpp'. 	* Fixed warnings in 'mdbadmin/opendb.cpp'. 	* Fixed warnings in 'mcallman/standard_age_report.cpp'. 	* Fixed warnings in 'mlm/mlm_airtools.c'. 	* Fixed warnings in 'mlm/mlm_remora.c'. 	* Fixed warnings in 'mlmh/mlm_functions.cpp'. 	* Fixed warnings in 'mlmh/vguest.cpp'. 	* Fixed warnings in 'mcs/mskimmer_widget.cpp'.

+++ libp/mlttydevice.cpp
@@ -590,7 +590,7 @@ void MLTTYDevice::writeTtyData()
 
   ioctl(tty_fd,TIOCOUTQ,&bytes);
   int n=2048-bytes;
-  if(tty_write_queue.size()<n) {
+  if((int)tty_write_queue.size()<n) {
     n=tty_write_queue.size();
   }
   if(n==0) {
